Add arrows -> as word separators for git diff-word
'-' and '>' characters are inverted in the regex because diff would
always show the dash otherwise.
Due to some regex dark magic, looks like the first separator after the
diff words is highlighted too. But not if it's the last in the regex.
Same thing still happens with slashes '/', but they are less common.